requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
django==5.0
django-environ==0.11.2
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-camel-case==1.4.2
gunicorn==21.2.0
django-cors-headers==4.3.1
black==22.3.0
whitenoise==6.6.0
django-better-admin-arrayfield==1.4.2
PyJWT==2.8.0
tonsdk==1.0.13